adding gaussian interp param to WIMT ( commented out )
also altering the meaning of prior weight in Atropos.

here, we use:

            RealType posteriorProbability = this->m_PriorProbabilityWeight *
              this->m_GaussianMixtureModelProportions[c] * prior
              + ( 1.0 - this->m_PriorProbabilityWeight ) *
              this->m_GaussianMixtureModelProportions[c] * likelihood * mrfPrior;

which allows us to tune between a FMM style solution and a likelihood solution.

not sure it's better, but produces better behavior in a crude example.

before we had (see picslblog) a mixture of mrf and likelihood on the first term.

RealType posteriorProbability = this->m_PriorProbabilityWeight *              this->m_GaussianMixtureModelProportions[c] * likelihood *              mrfPrior * prior + ( 1.0 - this->m_PriorProbabilityWeight ) *              this->m_GaussianMixtureModelProportions[c] * likelihood * mrfPrior;

we need to verify that the existing equation does not break old functionality.

can choose the new equation by setting: 

  bool  m_UseNewPosteriorEquation=true;

current default is :

  bool  m_UseNewPosteriorEquation=false;