COMP:  First attempt at adding SuperBuild
This add the superbuild system that has been
refined to accomplish several goals.  It is
more fully featured than is needed for the
simplest of ANTS superbuild systems.

The reason is that it very closely parallels
the superbuild system requirements that allow
this project to be added to the Slicer package,
and has been shown extensible and useful for the BRAINSTools
spharm-pdm and other packages.

Every attempt has been made to make this a generic
infrastructure that can be dropped on top of an existing
package, and with minimal editing add the superbuild system.

The differences from the BRAINSTools implementation include the
folowing:

1) Change LOCAL_PROJECT_NAME from "BRAINSTools" to "ANTS"
2) Remove custom BRAINSTools dependancy generation.