STYLE: Dupicate code inconsistency.
In an effort to track down a behavioral inconsistency
dupliate code was found in different compilation units
that was supposed to behave consistenty, but refactoring
had not been propogated to each copy of the function.

This effort lead to a code review that indicated many, many
places where redundant code had been copied and pasted.

In addtion, some header files had code that instantiated
the same objects in different compilation units.

All these issues were resolved by creating one common "antsUtilities"
static library for holding the common bits of code.  This can be
linked into executables or other shared libraries to produce products
with the same dependancies as is currently listed.