ENH: Propogate reference for warping consistently
When defining the output space based on an reference image,
it is better to use the SetOutputParametersFromImage function
in order to ensure that all elements of the space are
properly defined.

this->SetOutputOrigin ( image->GetOrigin() );
this->SetOutputSpacing ( image->GetSpacing() );
this->SetOutputDirection ( image->GetDirection() );
this->SetOutputStartIndex ( image->GetLargestPossibleRegion().GetIndex() );
this->SetSize ( image->GetLargestPossibleRegion().GetSize() );

In all cases, the output space StartIndex was not being set.
In a few cases, the SetOutputDirection was missed.

In addition, itk::WarpImageMultiTransformFilter was updated to
include and propgate the SetOutputParametersFromImage function along
with making the filter similar to that of the itk ResampleImage filter.