COMP: Tests for antsRegistration from BRAINSFit
A new suite of tests is being prepared. This
is the first patch set that adds new data
elements for the tests.

Currently TestSuite contains 10 test which are designed based on
BRAINSFit tests. More tests should be add later.

antsRegsitration tests are at a stable point now, and 10 new tests passed
successfully. More test will be added after as time permits.

The following tests are added and passed successfully:
1: antsRegistrationTest_AffineRotationMasks ......   Passed   35.96 sec
2: antsRegistrationTest_AffineScaleNoMasks ......... Passed   36.04 sec
3: antsRegistrationTest_AffineScaleMasks ........... Passed   36.15 sec
4: antsRegistrationTest_AffineTranslationNoMasks ... Passed   55.83 sec
5: antsRegistrationTest_RigidRotationHeadMasks.....  Passed   45.39 sec
6: antsRegistrationTest_RigidRotationNoMasks......   Passed   46.01 sec
7: antsRegistrationTest_RigidRotationMasks........   Passed   45.13 sec
8: antsRegistrationTest_AffineRotationNoMasks.....   Passed   30.77 sec
9: antsRegistrationTest_AffineTranslationMasks....   Passed   45.78 sec
10:antsRegistrationTest_RigidAnisotropicMasks.....   Passed  170.54 sec
100% tests passed, 0 tests
failed out of 10