BUG: Compiler warning indicated invalid code paths
There was both a template and a function with the same name.
The template was re-named to distiguish between the external
interface and the internal algorithmic implementation.

In addition, an unused parameter of the implementation
indicates that alternate code paths are not valid
due to lack of implementation for these alternate cases.
It appears that the implementation needs to be extended, so
for now an error message is printed if one of these unsupported
paths is attempted.