BUG: The interpolation behaviors were not consistent
THIS IS COMMON CODE FOUND THAT WAS SHARED BETWEEN TWO PROGRAMS
IT SHOULD BE A TEMPLATED FUNCTION.

I don't have time to do the complete refactoring at the moment, but
a function can soon be created to remove the need for the tmpl
file.

In merging the two programs, it was found that there were missing
and inconsistent functions between the two sets of copied and pasted
code.