BUG: Reimplement writting of separate name types when collapse-transforms is used.
There is a need to know if the Affine transform written to disk was an Affine transform that was
optimized, or an Affine transform that was collapsed from some other ::Linear type.  Unfortunatly, there
is not currently a way to know if an optimized Affine is collapsed to and Affine, and that would be written
out as a GenericAffine as well.