ENH: Add consistent similarity measure for tuning
When choosing parameters in the correct scope
for balancing speed and quality of registration, it
is very helpful to have a consistent view of the metric
value throughout the entire registration process.

Add "print-similarity-measure" option to antsRegistration
to measure the CC similarity metric between original fixed
image and original moving image at each itertion.