ENH: Allow user to specify the granularity of full size similarity testing
In order to evaluate how parameters affect the final registration results
it is important to have a consistent metric computed throughout the registration
process.  This ideal, however, is extraordinarily computationally expensive, so
the user is now allowed on the command line to specify how often (say every 15 iterations)
the full scale metric should be computed.