BUG: Add tests for CompositeTransformUtil tool
This patch provides a temporary fix for the bug in
'itkantsReadWriteTransform.h' file.
'CompositeTransformUtil' needs that the 'ReadTransform' function return
a Null transform when the input transform does not match to the input
dimension.
However, a static_cast statement is added to the last line of this
funciton so avoids a Null return. Although this added line can cause a bug
when we use the 'CompositeTransformUtil', it is needed by the ANTsR.

To make this code consistent with both the 'CompositeTransformUtil' and
'ANTsR', a new bool variable called "UsedStaticCastForR" is added as an
input to the 'ReadTransform' function that is defaulted to false.
If this variable is changed to true by the ANTsR, so an extra
lines is run at the end of the function that runs static_cast and possibly
changes the returned transform.

Also, one test is added to ANTS related to the CompositeTransformUtil tool.
This test makes sure that 'CompositeTransformUtil --assemble' is run
without any failure.
A second test will be added later that compares the output composite
transform of the 'CompositeTransformUtil --assemble' with a baseline composite
transform file to make sure that CompositeTransformUtil generates valid
results.

All the ANTS test are passed based on this patch.