ENH: Two more tests are added for CompositeTransformUtil
The "compareTwoCompositeTransforms" program is completed and works properly
now. Also, two more tests are added for the CompositeTransformUtil program.

The first test runs CompositeTransformUtil using "disassemble" flag and
compares the output displacement transform with the actual one that is
included in the baseline composite transform.

Then the created sub transforms by disassembling the baseline composite
transform are assembled again by CompositeTransformUtil and "assemble"
flag. The second test compares the output assembled composite transform
by the baseline.
This test uses the "compareTwoCompositeTransforms" program to do the
comparison.