WIP: Add 'float' option to ANTS for single precision computations
The final goal is to make "antsRegistration" program capable to
do its computations using single precision (float type) instead of
'double' precision.
It helps CPU to do more processings per time unit when the single precision is enough
for our computations.

By this patch antsRegistration is made as a template over computations type that can
accept 'float' or 'double'. The float precision can be enabled using the
"--float" flag. e.g. --float 1

Based on the new changes on ITK, its registration framework allows us to
do our computations in single precision, and the "float" option will
work when ANTs is built based on this new version of ITK.

After the changes of this pacth, ANTs is built successfully based on
the new version of itk.