COMP: Revert single precision transform & registration changes
This branch reverses the changes introduced by Ali Ghayoor that take
advantage of his post-4.4.0 ITK changes to implement single-precision
transforms and registration.

Note: this should NOT be merged.  This is meant to be a starting point
for a new ANTs checkpoint release.

It incorporates all current development changes, but by removing the
single-precision transform/registration stuff, it is compatible with ITK 4.4.0