BUG: Computation of brain volumes for CSV file failed
The script attempted to call $CORTICAL_THICKNESS_GM and $CORTICAL_THICKNESS_WM,
which at temp files that get deleted.

This caused the script to dump core, though it continues normally.

It is also arguably incorrect because $CORTICAL_THICKNESS_WM is the union of WM
(class 3) and deep gray (class 4).

The commands have been modified to use $BRAIN_SEGMENTATION_[GW]M.