Add option -y 0 to remove rigid components of the average affine transform when updating the template.
Made -y 1 the default for backwards compatibility with template construction scripts.

Hard coded -y 0 into antsLongitudinalCorticalThickness.sh. Because the number of images in this template construction call is small, the average translation / rotation can be substantial and this can cause the template to be rotated or shifted out of the field of view.

Also in antsLongitudinalCorticalThickness.sh, added a 5 voxel pad to the SST. If the top of the brain is cut off, or if there are varying amounts of spine / neck at the bottom of the images, padding the fixed image can improve affine alignment.