ENH: Update ITK for FixedParameter setting
ITK now preserves the fixed parameters (i.e. the center of rotation
for MatrixOffsetBase) objects during the inversion process.  This is important
so that the parameters of Inv(Inv(T)) ~= the parameters of T.  It is also
important if you desire to optimize the the inverse transform from the moving
image to the fixed image.