ENH: Make the collapsed transform keep the center of rotation
Previously the collapsed linear affine transform had fixed parameters
as all zero. Now we find the last transform that has a center, and set
that as the fixed parameters of the total transform, and it is done only
once.
If you wish to set the center explicitly, set before Offset computations
are done.  To define an affine transform, you must set the matrix, center,
and translation OR the matrix and offset