ENH: Improved organization and readability of write options
Added logic to better safegaurd users from mis-use of
the command line options.

When choosing to write composite transforms, do *NOT* write
the individual transform as well.  Choosing to write composite
transforms needs to be mutually exclusive from writing
individual transform files.  The reason is that the anticipated
behavior is different for these two options.

If writing the indiviual transforms, the user is required to know
that the existing initial-moving-transform needs to be incorporated
in the next "apply transform" step, so it is not included in the
writing process, (unless it was generated automatically with
geometry/centerofmass).  Composite transforms include the
initial-moving-transform implicitly.

This is further complicated by the new performance improvement
that initializes the current linear stage transform with the
previous stage output with implicit conversion.  This provides
performance improvements because there is only 1 linear transform
that needs to be used for computing jacobians/warping.