BUG: Fixed a bug in direct initialization per stage
The "InitializeWithPreviousLinearTransform" function should only return
"false" or true (NOT FAILURE or SUCCESS).
If direct intialization fails, the program should NOT be stopped because
the initial transform will be kept in the composite transform, and the
final results will be still correct.
A test is added to validate the "initialize-transforms-per-stage" option,
and a dependent test compares the outptut transform with a baseline.