BUG: antsMotionCorrDiffusionDirection would silently produce wrong output if you used the DWI or other 4D image as a reference.
This is because itkImageFileReader will set the direction matrix to identity if it is reading a 4D image with a reader templated as 3D.

The general use case for antsMotionCorr in DTI is correction to a fixed image (eg average b=0 image). This 3D image should be used as the reference image for antsMotionCorrDiffusionDirection.