ENH: allow control of atropos internal iterations
we are running out of letters for the act command line.  i prefer to use 25 or more iterations
for difficult segmentation tasks where the priors may be very smooth and may take longer to
converge to a solution.  this change allows, from the antsCorticalThickness command line, the
user to alter this value as he or she wishes.