ENH:  Normalize output to input dynamic range
When applying a denoising filter, it is natural
to think that the noise is removed and that the
resulting image would have a range very similar
to the input image.  In order to accomplish this
the output image is linearly rescaled to the dynamic
range of the input image.

This has the effect of making the error image have
the smallest magnitude (in terms of sum of square error)
and is easier to visualize and interpret.