ENH: Masking operations uniform use nonzero values
When applying a masking operation, any non-zero value
in the image should be considered foreground.  This is
the behavior of the itkMaskImageFilter, and the
itkSpatialImageMaskObject filters.  This makes using
the same image (even an label map image with multiple
labels) appropriate across different uses (i.e. registration
and segmentation).

In all uses inside ANTs, this had been hard coded to the
forground being equal to a value of 1, and no cases were found
that allowed for changing the mask match value.  For masks that
have only values of zero and 1, backwards compatibility is preserved.

For uses where label maps or intensity images were used, this
changes behavior, but in a way that makes ants tools consistent with
other ITK based behaviors.