WIP: Change tensor / vector treatment in antsApplyTransforms
Now assume that tensor is based in the moving voxel space, and reorient
to the fixed voxel space before resampling. Same with vectors.

This fixes a bug where the tensors were transformed in the opposite direction.
RebaseTensorImage had the same error but only when the REFERENCE option was used.

To do:

 - More tests. There's new oblique Prisma data on the dcm2niix page