Add support for an --interpolation flag in antsMotionCorr
Note that no support was added for either Gaussian or genericLabel
interpolation routines, as those require adapting parameters according
to the voxel dimensions. If users want these, they can always fall
back to using antsApplyTransforms for a more generic solution.

Note that the short name for --interpolation is -p because -n is already
in use.