ENH: Added bash script for ANTs tissue segmentation
ANTs Atropos and antsAtroposN4.sh need all the priors already warped to the subject space before the segmentation is run. This developed script combines the registration step with antsAtroposN4 run.

This script runs:
- brain extraction
- template registration
- atropos segmentation
- (optional) register to template space
- QC (mosaic QC image)

This script was adopted and modified from antsCorticalThickness.sh,
so it follows similar command line convention.

This script was tested on several arbitrarily selected
subject/templates, and it ran successfully and generated the requested outputs.