ENH: Version info from a config file for releases
Did a bunch of refactoring of version handling, consolidating the work
into ProjectSourceVersion.cmake (handling the git side and setting
_GIT_VERSION_MAJOR etc) and Version.cmake (deciding the final form of
the version info).

The version string should now match git-describe more closely. The old
behavior was to look for commits since Version.cmake was changed, but
since that is entirely auto-populated, it is not a very useful metric.
Now use post-X, where X is the number of commits since the last
release.

For releases, we need to provide a file
${CMAKE_CURRENT_SOURCE_DIR}/CMake/ReleaseSourceVersionVars.cmake
which should contain, eg:

set( _GIT_VERSION_MAJOR "2" )
set( _GIT_VERSION_MINOR "3" )
set( _GIT_VERSION_PATCH "6" )