ENH: apply transform to single slice of time series
A common use case is to register a time series to a scalar reference,
then apply the warp to each volume. Sometimes, users will have different
transforms for each volume (say, motion correction in fMRI). Other times,
the time series might be very large after resampling into the fixed space.

This commit adds an option --time-index, which allows the user to select
a particular volume. This is then extracted and transformed as a scalar image.