ENH: Functions to convert to / from FSL tensor
FSL requires radiological orientation of bvecs / tensors, regardless
of image orientation. This causes a problem for "NEUROLOGICAL" images
(positive determinant of direction matrix), because the tensors are then
not transformed to physical space by the direction matrix.

On import, we fix this by applying a reflection in x, and we do the same
on export back to FSL tensor format.